A wide range of changes here:
Replaced usages of URL with URI, since URL.hashCode() does a DNS lookup for no good reason
This involved renaming some classes, methods, database tables, etc, from url to uri
Implemented a URI Tree model, so we can show the URI's that we have seen.
Updated the SummaryView to show the URI tree as well as the Conversation Table.
Redid the about.txt file. Even though it says txt, it should be HTML formatted. Note that renaming it to .html doesn't work, since Eclipse will not copy the file to the build dir. Doh!
Reimplemented the JDBC DAO's, so that we do not hard code the database type at build/run-time. Now, to add support for new databases, or to change the CREATE TABLE statements, simply modify the database.properties file