Added support for annotations. These allow the user/operator to describe a conversation, for example:
"Clicked on the login button"

They are not currently visible, but will be at some time in the future.

There is also a new ProxyToolBar, which provides a "stays on top" widget designed to
float in the browser's title bar (or wherever you want to place it), conveniently located
for the operator to describe what he is about to do.

There is also initial support for "Findings". This is separate to the Annotations, in that
the Findings are attached to potentially multiple conversations, and specifically describe
something that should be reported. There is no current use of these classes, but the
envisaged use will be something like:

Select a number of conversations that make up your finding.
Right-click and choose "Add finding".
This will display a text box where you can enter the description of the finding, (potentially
modify the related conversations), and then "Save" the finding. The Finding could later be
exported using AVDL, or another suitable reporting tool.