HttpService: Move routines for actually fetching responses into a single place
It makes sense to have all the routines for fetching responses in a single place,
so that we can configure proxies, etc simply, without forcing plugins to restart.
It also means that we can track which clients are still 'live', i.e. may still be
used by a Thread, so we can clean up. This is not yet implemented.