Move to eliminate ConversationSummary
This continues the move to eliminate ConversationSummary, by making
the Conversation classes more self-contained.

Now Conversation is effectively a Transfer Object, that can be
easily extended to allow for a cached version, backed by a DAO
for getting the larger portions of the request/response (ie. the body)

We introduce a StreamingConversation that includes responseContentStream
functionality moved from Conversation, and improve BufferedConversation
at the same time.