MAJOR changes: We no longer use ConversationSummary for anything
This is important in that we are now able to add the entire Conversation
to the list of Conversations shown to the user. Any operations on the
Conversation will have all the details of the Conversation at its
disposal, rather than having to indirect via the ConversationSummary.

For memory conservation purposes, we implement a "cached" Conversation,
that can look up "expensive" details such as its headers and contents
from the database tables.