Upgrade Spring Framework to 2.0.1
This solves the problem where one saw "commandLabel" instead of the
correct labels for menu items.

This also specifies exclusions in the automatic dependencies, which
help to reduce the size of the application download (e.g. AOP-related
libraries), as well as excluding empty jars that JavaWS 1.6 had issues
verifying. Apparently it will not accept an empty jar as being signed,
whereas 1.5 was quite happy with this.