ParsedResponseForm: make it editable
This updates ParsedRequestForm to behave in a similar way to the
ParsedRequestForm. This is largely a copy and paste, with suitable
renaming. The code dealing with URL parameter nodes and cookie nodes
was removed. It may be desirable to present Set-Cookie headers in
a similar expanded way, although this can be somewhat non-intuitive, too.