Make SiteMapView only show "successful" conversations
I.e. only show conversations with a response status
starting with 2xx or 3xx.

Also make sure that we only remove items from the tree when
the last instance has been removed from the list. This will
be necessary when we filter conversations based on whether they
are images or not.